Add javadoc to BidRestController, add logs to BidListService

+++ Poseiden-skeleton/src/main/java/com/nnk/springboot/services/BidListService.java
@@ -1,10 +1,12 @@
 package com.nnk.springboot.services;
 
 import com.nnk.springboot.domain.BidList;
+import com.nnk.springboot.exceptions.ResourceNotFoundException;
 import com.nnk.springboot.repositories.BidListRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.util.Assert;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -25,7 +27,6 @@ public class BidListService {
      * Creates new bid.
      */
     public BidList saveBid(final BidList bidList) {
-
         bidList.setCreationDate(LocalDateTime.now());
         return bidListRepository.save(bidList);
     }
@@ -39,6 +40,7 @@ public class BidListService {
      * @return Optional bid list object.
      */
     public Optional<BidList> getBidById(final Integer id) {
+        Assert.notNull(id, "ID must not be null.");
         return bidListRepository.findById(id);
     }
 
@@ -58,8 +60,13 @@ public class BidListService {
      * @return updated bid object.
      */
     public BidList updateBid(final BidList bidList) {
-        bidList.setRevisionDate(LocalDateTime.now());
-        return bidListRepository.save(bidList);
+        if (bidListRepository.findById(bidList.getBidListId()).isPresent()) {
+            bidList.setRevisionDate(LocalDateTime.now());
+            return bidListRepository.save(bidList);
+        } else {
+            log.error("Provided bid with ID " + bidList.getBidListId() + "does not exist.");
+            throw new ResourceNotFoundException("Provided bid with ID " + bidList.getBidListId() + "does not exist.");
+        }
     }
 
     /**
@@ -69,7 +76,14 @@ public class BidListService {
      *         ID of bid to delete.
      */
     public void deleteBid(final Integer id) {
-        bidListRepository.deleteById(id);
+        Assert.notNull(id, "ID must not be null.");
+        if (bidListRepository.findById(id).isPresent()) {
+            bidListRepository.deleteById(id);
+            log.debug("Deleted bid with ID "+ id +".");
+        } else {
+            log.error("There is no such bid with ID " + id + ".");
+            throw new ResourceNotFoundException("There is no such bid with ID " + id + ".");
+        }
     }
 
 }
